backend/requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
faiss-cpu

langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.25
langchain-text-splitters

openai